refactor(web-ui): decompose Today.tsx + tab Universes detail
Today.tsx (1196→~220 lines): extract TodayActionItems and
TodayActionList into components/domain/today/. Page file keeps
only the outer layout and small widgets (WeeklyReviewNudge,
TodayPrioritySection, PendingOrdersBadge).

Universes.tsx: collapse flat detail + top-level LiveDiscovery card
into four tabs (Config / Constituents / Discovery / Screener).
"Run Screener" auto-advances to the Screener tab.
